requirements.txt
fastapi
uvicorn
python-dotenv
openai==0.28
pydantic

# python-multipart
